requirements.txt
flask
requests
beautifulsoup4
dnspython
